+++ modules/control-utility/src/main/java/org/apache/ignite/internal/commandline/CommandHandler.java
@@ -17,12 +17,10 @@
 
 package org.apache.ignite.internal.commandline;
 
-import java.lang.reflect.Field;
 import java.net.URL;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.LinkedHashMap;
@@ -31,11 +29,6 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Scanner;
 import java.util.UUID;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.BiConsumer;
-import java.util.function.Consumer;
 import javax.cache.configuration.Factory;
 import javax.net.ssl.SSLContext;
 import org.apache.ignite.IgniteCheckedException;
@@ -49,20 +42,12 @@ import org.apache.ignite.internal.client.GridClientDisconnectedException;
 import org.apache.ignite.internal.client.GridClientHandshakeException;
 import org.apache.ignite.internal.client.GridServerUnreachableException;
 import org.apache.ignite.internal.client.impl.connection.GridClientConnectionResetException;
+import org.apache.ignite.internal.dto.IgniteDataTransferObject;
 import org.apache.ignite.internal.logger.IgniteLoggerEx;
 import org.apache.ignite.internal.management.IgniteCommandRegistry;
-import org.apache.ignite.internal.management.api.Argument;
 import org.apache.ignite.internal.management.api.BeforeNodeStartCommand;
-import org.apache.ignite.internal.management.api.CliPositionalSubcommands;
 import org.apache.ignite.internal.management.api.Command;
-import org.apache.ignite.internal.management.api.CommandUtils;
-import org.apache.ignite.internal.management.api.CommandsRegistry;
-import org.apache.ignite.internal.management.api.ComputeCommand;
-import org.apache.ignite.internal.management.api.EnumDescription;
 import org.apache.ignite.internal.management.api.HelpCommand;
-import org.apache.ignite.internal.management.api.LocalCommand;
-import org.apache.ignite.internal.management.api.Positional;
-import org.apache.ignite.internal.management.api.WithCliConfirmParameter;
 import org.apache.ignite.internal.management.cache.CacheCommand;
 import org.apache.ignite.internal.util.IgniteUtils;
 import org.apache.ignite.internal.util.spring.IgniteSpringHelperImpl;
@@ -76,6 +61,7 @@ import org.apache.ignite.plugin.security.SecurityCredentialsProvider;
 import org.apache.ignite.ssl.SslContextFactory;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.context.ApplicationContext;
+
 import static java.lang.System.lineSeparator;
 import static java.util.Objects.nonNull;
 import static org.apache.ignite.IgniteSystemProperties.IGNITE_ENABLE_EXPERIMENTAL_COMMAND;
@@ -85,18 +71,12 @@ import static org.apache.ignite.internal.commandline.CommandLogger.errorMessage;
 import static org.apache.ignite.internal.commandline.CommonArgParser.CMD_AUTO_CONFIRMATION;
 import static org.apache.ignite.internal.commandline.CommonArgParser.CMD_ENABLE_EXPERIMENTAL;
 import static org.apache.ignite.internal.commandline.CommonArgParser.CMD_VERBOSE;
-import static org.apache.ignite.internal.management.AbstractCommandInvoker.visitCommandParams;
-import static org.apache.ignite.internal.management.api.CommandUtils.CMD_WORDS_DELIM;
 import static org.apache.ignite.internal.management.api.CommandUtils.DOUBLE_INDENT;
 import static org.apache.ignite.internal.management.api.CommandUtils.INDENT;
-import static org.apache.ignite.internal.management.api.CommandUtils.PARAMETER_PREFIX;
-import static org.apache.ignite.internal.management.api.CommandUtils.PARAM_WORDS_DELIM;
 import static org.apache.ignite.internal.management.api.CommandUtils.asOptional;
 import static org.apache.ignite.internal.management.api.CommandUtils.cmdText;
 import static org.apache.ignite.internal.management.api.CommandUtils.join;
-import static org.apache.ignite.internal.management.api.CommandUtils.parameterExample;
 import static org.apache.ignite.internal.management.api.CommandUtils.toFormattedCommandName;
-import static org.apache.ignite.internal.management.api.CommandUtils.valueExample;
 
 /**
  * Class that execute several commands passed via command line.
@@ -235,7 +215,7 @@ public class CommandHandler {
      * @param rawArgs Arguments to parse and execute.
      * @return Exit code.
      */
-    public int execute(List<String> rawArgs) {
+    public <A extends IgniteDataTransferObject> int execute(List<String> rawArgs) {
         LocalDateTime startTime = LocalDateTime.now();
 
         Thread.currentThread().setName("session=" + ses);
@@ -259,10 +239,11 @@ public class CommandHandler {
 
             verbose = F.exist(rawArgs, CMD_VERBOSE::equalsIgnoreCase);
 
-            ConnectionAndSslParameters args = new CommonArgParser(logger, cmds).parseAndValidate(rawArgs.iterator());
+            ConnectionAndSslParameters<A> args = new CommonArgParser(logger, cmds).parseAndValidate(rawArgs);
 
-            DeclarativeCommandAdapter<?> command = args.declarativeCmd();
-            commandName = command.name();
+            CommandInvoker<A> invoker = new CommandInvoker<>(args.command(), args.commandArg());
+
+            commandName = toFormattedCommandName(args.command().getClass()).toUpperCase();
 
             GridClientConfiguration clientCfg = getClientConfiguration(args);
 
@@ -274,10 +255,11 @@ public class CommandHandler {
 
             while (true) {
                 try {
-                    if (!args.autoConfirmation()) {
-                        command.prepareConfirmation(clientCfg, logger);
+                    if (!invoker.prepare(clientCfg, logger))
+                        return EXIT_CODE_OK;
 
-                        if (!confirm(command.confirmationPrompt())) {
+                    if (!args.autoConfirmation()) {
+                        if (!confirm(invoker.confirmationPrompt())) {
                             logger.info("Operation cancelled.");
 
                             return EXIT_CODE_OK;
@@ -288,10 +270,12 @@ public class CommandHandler {
                     logger.info("Arguments: " + argumentsToString(rawArgs));
                     logger.info(U.DELIM);
 
-                    if (command.isHelp())
-                        printUsage(logger, command.command());
+                    if (args.command() instanceof HelpCommand)
+                        printUsage(logger, args.command());
+                    else if (args.command() instanceof BeforeNodeStartCommand)
+                        lastOperationRes = invoker.invokeBeforeNodeStart(clientCfg, logger);
                     else
-                        lastOperationRes = command.execute(clientCfg, logger, args.verbose());
+                        lastOperationRes = invoker.invoke(clientCfg, logger, args.verbose());
 
                     break;
                 }
@@ -790,6 +774,21 @@ public class CommandHandler {
         logger.info(DOUBLE_INDENT + EXIT_CODE_UNEXPECTED_ERROR + " - unexpected error.");
     }
 
+    /**
+     * Print info for user about command (parameters, use cases and so on).
+     *
+     * @param logger Logger to use.
+     */
+    public void printUsage(IgniteLogger logger, Command<?, ?> cmd) {
+        if (cmd instanceof CacheCommand || cmd instanceof CacheCommand.CacheHelpCommand)
+            printCacheHelpHeader(logger);
+
+        new CommandInvoker<>(cmd, null).usage(logger);
+
+        if (cmd instanceof CacheCommand || cmd instanceof CacheCommand.CacheHelpCommand)
+            logger.info("");
+    }
+
     /** */
     private void printCacheHelpHeader(IgniteLogger logger) {
         logger.info(INDENT + "The '--cache subcommand' is used to get information about and perform actions" +
@@ -804,190 +803,5 @@ public class CommandHandler {
         logger.info("");
         logger.info("");
         logger.info(INDENT + "Subcommands:");
-    }
-
-    /**
-     * Print info for user about command (parameters, use cases and so on).
-     *
-     * @param logger Logger to use.
-     */
-    public void printUsage(IgniteLogger logger, Command<?, ?> cmd) {
-        if (cmd instanceof CacheCommand || cmd instanceof CacheCommand.CacheHelpCommand)
-            printCacheHelpHeader(logger);
-
-        usage(cmd, Collections.emptyList(), logger);
-
-        if (cmd instanceof CacheCommand || cmd instanceof CacheCommand.CacheHelpCommand)
-            logger.info("");
-    }
-
-    /**
-     * Generates usage for base command and all of its children, if any.
-     *
-     * @param cmd Base command.
-     * @param parents Collection of parent commands.
-     * @param logger Logger to print help to.
-     */
-    private void usage(Command<?, ?> cmd, List<Command<?, ?>> parents, IgniteLogger logger) {
-        if (cmd instanceof LocalCommand
-            || cmd instanceof ComputeCommand
-            || cmd instanceof HelpCommand
-            || cmd instanceof BeforeNodeStartCommand) {
-            logger.info("");
-
-            if (cmd.experimental())
-                logger.info(INDENT + "[EXPERIMENTAL]");
-
-            printExample(cmd, parents, logger);
-
-            if (CommandUtils.hasDescribedParameters(cmd)) {
-                logger.info("");
-                logger.info(DOUBLE_INDENT + "Parameters:");
-
-                AtomicInteger maxParamLen = new AtomicInteger();
-
-                Consumer<Field> lenCalc = fld -> {
-                    maxParamLen.set(Math.max(maxParamLen.get(), parameterExample(fld, false).length()));
-
-                    if (fld.isAnnotationPresent(EnumDescription.class)) {
-                        EnumDescription enumDesc = fld.getAnnotation(EnumDescription.class);
-
-                        for (String name : enumDesc.names())
-                            maxParamLen.set(Math.max(maxParamLen.get(), name.length()));
-                    }
-                };
-
-                visitCommandParams(cmd.argClass(), lenCalc, lenCalc, (argGrp, flds) -> flds.forEach(lenCalc));
-
-                Consumer<Field> printer = fld -> {
-                    BiConsumer<String, String> logParam = (name, description) -> {
-                        if (F.isEmpty(description))
-                            return;
-
-                        logger.info(
-                            DOUBLE_INDENT + INDENT + U.extendToLen(name, maxParamLen.get()) + "  - " + description + "."
-                        );
-                    };
-
-                    if (!fld.isAnnotationPresent(EnumDescription.class)) {
-                        logParam.accept(
-                            parameterExample(fld, false),
-                            fld.getAnnotation(Argument.class).description()
-                        );
-                    }
-                    else {
-                        EnumDescription enumDesc = fld.getAnnotation(EnumDescription.class);
-
-                        String[] names = enumDesc.names();
-                        String[] descriptions = enumDesc.descriptions();
-
-                        for (int i = 0; i < names.length; i++)
-                            logParam.accept(names[i], descriptions[i]);
-                    }
-                };
-
-                visitCommandParams(cmd.argClass(), printer, printer, (argGrp, flds) -> {
-                    flds.stream().filter(fld -> fld.isAnnotationPresent(Positional.class)).forEach(printer);
-                    flds.stream().filter(fld -> !fld.isAnnotationPresent(Positional.class)).forEach(printer);
-                });
-            }
-        }
-
-        if (cmd instanceof CommandsRegistry) {
-            List<Command<?, ?>> parents0 = new ArrayList<>(parents);
-
-            parents0.add(cmd);
-
-            ((CommandsRegistry<?, ?>)cmd).commands().forEachRemaining(cmd0 -> usage(cmd0.getValue(), parents0, logger));
-        }
-    }
-
-    /**
-     * Generates and prints example of command.
-     *
-     * @param cmd Command.
-     * @param parents Collection of parent commands.
-     * @param logger Logger to print help to.
-     */
-    private void printExample(Command<?, ?> cmd, List<Command<?, ?>> parents, IgniteLogger logger) {
-        logger.info(INDENT + cmd.description() + ":");
-
-        StringBuilder bldr = new StringBuilder(DOUBLE_INDENT + UTILITY_NAME);
-
-        AtomicBoolean prefixInclude = new AtomicBoolean(true);
-
-        AtomicReference<String> parentPrefix = new AtomicReference<>();
-
-        Consumer<Object> namePrinter = cmd0 -> {
-            bldr.append(' ');
-
-            if (prefixInclude.get())
-                bldr.append(PARAMETER_PREFIX);
-
-            String cmdName = toFormattedCommandName(cmd0.getClass());
-
-            String parentPrefix0 = parentPrefix.get();
-
-            parentPrefix.set(cmdName);
-
-            if (!F.isEmpty(parentPrefix0)) {
-                cmdName = cmdName.replaceFirst(parentPrefix0 + CMD_WORDS_DELIM, "");
-
-                if (!prefixInclude.get())
-                    cmdName = cmdName.replaceAll(CMD_WORDS_DELIM + "", PARAM_WORDS_DELIM + "");
-            }
-
-            bldr.append(cmdName);
-
-            if (cmd0 instanceof CommandsRegistry)
-                prefixInclude.set(!(cmd0.getClass().isAnnotationPresent(CliPositionalSubcommands.class)));
-        };
-
-        parents.forEach(namePrinter);
-        namePrinter.accept(cmd);
-
-        BiConsumer<Boolean, Field> paramPrinter = (spaceReq, fld) -> {
-            if (spaceReq)
-                bldr.append(' ');
-
-            bldr.append(parameterExample(fld, true));
-        };
-
-        visitCommandParams(
-            cmd.argClass(),
-            fld -> bldr.append(' ').append(valueExample(fld)),
-            fld -> paramPrinter.accept(true, fld),
-            (argGrp, flds) -> {
-                if (argGrp.onlyOneOf()) {
-                    bldr.append(' ');
-
-                    if (argGrp.optional())
-                        bldr.append('[');
-
-                    for (int i = 0; i < flds.size(); i++) {
-                        if (i != 0)
-                            bldr.append('|');
-
-                        paramPrinter.accept(false, flds.get(i));
-                    }
-
-                    if (argGrp.optional())
-                        bldr.append(']');
-                }
-                else {
-                    flds.stream()
-                        .filter(fld -> fld.isAnnotationPresent(Positional.class))
-                        .forEach(fld -> bldr.append(' ').append(valueExample(fld)));
-                    flds.stream()
-                        .filter(fld -> !fld.isAnnotationPresent(Positional.class))
-                        .forEach(fld -> paramPrinter.accept(true, fld));
-                }
-            }
-        );
-
-        if (cmd.argClass().isAnnotationPresent(WithCliConfirmParameter.class))
-            bldr.append(' ').append(CommandUtils.asOptional(CMD_AUTO_CONFIRMATION, true));
-
-        logger.info(bldr.toString());
     }
 }
