+++ modules/core/src/main/java/org/apache/ignite/internal/management/ChangeTagCommand.java
@@ -18,7 +18,6 @@
 package org.apache.ignite.internal.management;
 
 import java.util.function.Consumer;
-import org.apache.ignite.internal.client.GridClient;
 import org.apache.ignite.internal.management.api.ComputeCommand;
 import org.apache.ignite.internal.visor.misc.VisorClusterChangeTagTask;
 import org.apache.ignite.internal.visor.misc.VisorClusterChangeTagTaskResult;
@@ -54,7 +53,7 @@ public class ChangeTagCommand implements ComputeCommand<ChangeTagCommandArg, Vis
     }
 
     /** {@inheritDoc} */
-    @Override public @Nullable String confirmationPrompt(GridClient cli, ChangeTagCommandArg arg) {
+    @Override public @Nullable String confirmationPrompt(ChangeTagCommandArg arg) {
         return "Warning: the command will change cluster tag.";
     }
 }
