+++ modules/core/src/main/java/org/apache/ignite/internal/management/cdc/CdcDeleteLostSegmentLinksCommand.java
@@ -22,12 +22,12 @@ import java.util.Collections;
 import java.util.Map;
 import java.util.UUID;
 import java.util.function.Consumer;
-import org.apache.ignite.internal.client.GridClient;
 import org.apache.ignite.internal.management.api.ComputeCommand;
 import org.apache.ignite.internal.management.api.ExperimentalCommand;
 import org.apache.ignite.internal.util.typedef.T3;
 import org.apache.ignite.internal.util.typedef.internal.U;
 import org.apache.ignite.internal.visor.cdc.VisorCdcDeleteLostSegmentsTask;
+
 import static org.apache.ignite.internal.management.api.CommandUtils.servers;
 
 /**
@@ -67,7 +67,7 @@ public class CdcDeleteLostSegmentLinksCommand implements
     }
 
     /** {@inheritDoc} */
-    @Override public String confirmationPrompt(GridClient cli, CdcDeleteLostSegmentLinksCommandArg arg) {
+    @Override public String confirmationPrompt(CdcDeleteLostSegmentLinksCommandArg arg) {
         return "Warning: The command will fix WAL segments gap in case CDC link creation was stopped by distributed " +
             "property or excess of maximum CDC directory size. Gap will be fixed by deletion of WAL segment links" +
             "previous to the last gap." + U.nl() +
