+++ modules/core/src/main/java/org/apache/ignite/internal/management/encryption/EncryptionChangeMasterKeyCommand.java
@@ -18,7 +18,6 @@
 package org.apache.ignite.internal.management.encryption;
 
 import java.util.function.Consumer;
-import org.apache.ignite.internal.client.GridClient;
 import org.apache.ignite.internal.management.api.ComputeCommand;
 import org.apache.ignite.internal.visor.encryption.VisorChangeMasterKeyTask;
 
@@ -45,7 +44,7 @@ public class EncryptionChangeMasterKeyCommand implements ComputeCommand<Encrypti
     }
 
     /** {@inheritDoc} */
-    @Override public String confirmationPrompt(GridClient cli, EncryptionChangeMasterKeyCommandArg arg) {
+    @Override public String confirmationPrompt(EncryptionChangeMasterKeyCommandArg arg) {
         return "Warning: the command will change the master key. Cache start and node join during the key change " +
             "process is prohibited and will be rejected.";
     }
