+++ modules/core/src/main/java/org/apache/ignite/internal/management/meta/MetaUpdateCommand.java
@@ -19,7 +19,6 @@ package org.apache.ignite.internal.management.meta;
 
 import java.util.function.Consumer;
 import org.apache.ignite.internal.binary.BinaryMetadata;
-import org.apache.ignite.internal.client.GridClient;
 import org.apache.ignite.internal.commandline.meta.tasks.MetadataMarshalled;
 import org.apache.ignite.internal.commandline.meta.tasks.MetadataUpdateTask;
 import org.apache.ignite.internal.management.api.ComputeCommand;
@@ -59,7 +58,7 @@ public class MetaUpdateCommand implements
     }
 
     /** {@inheritDoc} */
-    @Override public @Nullable String confirmationPrompt(GridClient cli, MetaUpdateCommandArg arg) {
+    @Override public @Nullable String confirmationPrompt(MetaUpdateCommandArg arg) {
         return "Warning: the command will update the binary metadata at the cluster.";
     }
 }
